+++ clc-engine/src/expression/macros.rs
@@ -10,7 +10,7 @@ macro_rules! cst {
 /// cst_exp!("e")
 macro_rules! cst_exp {
     ($c:expr) => {
-        crate::expression::Expression::constant(crate::macros::cst!($c))
+        crate::expression::Expression::constant(crate::expression::macros::cst!($c))
     };
 }
 
@@ -22,21 +22,44 @@ macro_rules! lit {
     };
 }
 
+/// Return Variable expression.
+/// var!("x").
+macro_rules! var {
+    ($ident:expr) => {
+        crate::expression::Variable::new($ident)
+    };
+}
+
+/// Return lambda expression.
+/// lambda!(["x"], node!("x", '^', 2))
+macro_rules! lambda {
+    ([  $var:expr ], $expr:expr) => {
+        crate::expression::Lambda::new($var, $expr)
+    };
+}
+
 /// Return function call expression.
 /// fc_exp!("sqrt", 100)
 macro_rules! fc_exp {
-        ($ident:expr, $( $expr:expr ),+) => {{
-            let fc = crate::macros::fc!($ident, $( $expr ),+ );
+        ($ident:expr, [ $( $expr:expr ),+ ], $lambda:expr) => {{
+            let fc = crate::expression::macros::fc!($ident, [ $( $expr ),+ ] , $lambda);
             crate::expression::Expression::func_call(fc)
-        }}
+        }};
+        ($ident:expr, $( $expr:expr ),+) => {{
+            let fc = crate::expression::macros::fc!($ident, $( $expr ),+ );
+            crate::expression::Expression::func_call(fc)
+        }};
 }
 
 /// Return function call.
 /// fc!("sqrt", 100)
 macro_rules! fc {
-        ($ident:expr, $( $expr:expr ),+) => {{
-           crate::expression::FuncCall::new($ident, vec![ $( $expr ),+ ])
-        }}
+        ($ident:expr, [ $( $expr:expr ),+], $lambda:expr ) => {
+            crate::expression::FuncCall::new($ident, vec![ $( $expr ),+ ], Some($lambda))
+        };
+        ($ident:expr, $( $expr:expr ),+) => {
+           crate::expression::FuncCall::new($ident, vec![ $( $expr ),+ ], None)
+        };
 }
 
 /// Return node expression.
@@ -55,5 +78,7 @@ pub(crate) use cst;
 pub(crate) use cst_exp;
 pub(crate) use fc;
 pub(crate) use fc_exp;
+pub(crate) use lambda;
 pub(crate) use lit;
 pub(crate) use node;
+pub(crate) use var;