+++ clc-engine/src/lib.rs
@@ -29,8 +29,6 @@ mod calculator;
 mod errors;
 mod eval;
 mod expression;
-#[cfg(test)]
-mod macros;
 mod parse;
 
 pub use {